requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
djangorestframework-simplejwt>=5.0,<6.0
psycopg2-binary
redis
django-redis
drf-yasg
django-cors-headers 
django-debug-toolbar  
django-environ       
Pillow              
pytest              
pytest-django        
